requirements.txt
certifi==2017.1.23
click==6.7
cx-Oracle==5.2
Flask==0.12.2
Flask-Cors==3.0.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pyaml==16.12.2
PyYAML==3.12
six==1.10.0
Werkzeug==0.12.2
wincertstore==0.2
